requirements.txt
flask
flask-cors
waitress
python-dotenv
Flask-SQLAlchemy
passlib